requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
psutil>=5.9.0
